requirements.txt
flask
flask-cors
ariadne
graphql-core>=3.0.0
SQLAlchemy
clerk-backend-api
python-dotenv
bson
pymysql
flask-socketio
eventlet